dom/html/test/forms/test_valueAsDate_pref.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=874640
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 874640</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 874640 **/
  var states = [
  // dom.experimental_forms, dom.forms.datepicker, dom.forms.datetime, expectedValueAsDate
  [ 'true', 'true', 'true', 'true' ],
  [ 'true', 'false', 'false', 'true' ],
  [ 'false', 'true', 'false', 'true' ],
  [ 'false', 'false', 'true', 'true' ],
  [ 'false', 'false', 'false', 'false' ],
        'end'
  ];

  SimpleTest.waitForExplicitFinish();

  function runTest(iframe) {
    var state = states.shift();

    if (state == 'end') {
      SimpleTest.finish();
      return;
    }

    SpecialPowers.pushPrefEnv({"set":[
      ["dom.experimental_forms", state[0] === 'true'],
      ["dom.forms.datepicker", state[1] === 'true'],
      ["dom.forms.datetime", state[2] === 'true']]},
      function() {
        iframe.src = 'data:text/html,<script>' +
                 'parent.is("valueAsDate" in document.createElement("input"), ' +
                 state[3] + ', "valueAsDate presence state should be ' + state[3] + '");' +
                 '<\/script>'
     });
  }

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=874640">Mozilla Bug 874640</a>
<p id="display"></p>
<div id="content" style="display: none">
  <iframe onload='runTest(this);'></iframe>
</div>
<pre id="test">
</pre>
</body>
</html>
